[CASSANDRA-20476] Relocate Discovery to new package
